test_utils/virt: Move VM-agnostic functions
to a new `NodeBackend` trait, namely, `ssh_exec`, `get_root_key`, and
KubeVirt's implementation of `verify_encrypted_root` (which Azure
overrides), for sharing in node-level tests.
